refactor(compare): extract C9 to checks/clinical_ranges.py (Phase A 10/N)
Move check_c9_clinical_range and its _range_violations helper into the
checks subpackage. clinical_ranges.py needs `re` for word-boundary
substring matching of range names.

The `_C10_SIMPLE_RE` constant (which was defined immediately after C9 in
the original layout) belongs with C10 — it's been moved back into the
main file just above check_c10_cross_variable to keep it there until
C10 extraction.

Regression-check: COPDGene, ARIC, HCHS — MD + JSON all match. Tests:
79/79 pass.

Main file: 1051 → 931 lines.
